Improve compilation time

+++ include/utils.hpp
@@ -1,8 +1,7 @@
 #ifndef __UTILS_HPP_
 #define __UTILS_HPP_
 
-#include <string>
-#include <filesystem>
+#include <thread>
 #include <vector>
 
 std::vector<uint8_t> read_file(const char* path);
@@ -16,7 +15,7 @@ inline float deg_to_rad(float deg) {
 }
 
 inline float randd() {
-    return rand() / (RAND_MAX + 1.f);
+    return (float)rand() / (RAND_MAX + 1.f);
 }
 
 inline float randd(float min, float max) {
@@ -24,7 +23,7 @@ inline float randd(float min, float max) {
 }
 
 inline int32_t randi(int32_t min, int32_t max) {
-    return (int32_t)(randd(min, max + 1.f));
+    return (int32_t)(randd((float)min, (float)max + 1.f));
 }
 
 inline float clamp(float x, float min, float max) {
